+++ crates/shielder-relayer/src/lib.rs
@@ -1,18 +1,16 @@
-use std::str::FromStr;
-
 use axum::{
     http::StatusCode,
     response::{IntoResponse, Response},
     Json,
 };
-use clap::ValueEnum;
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use shielder_contract::alloy_primitives::{Address, Bytes, FixedBytes, TxHash, U256};
-use strum_macros::EnumIter;
 
 mod environment_variables;
 pub use environment_variables::*;
+mod currency;
+pub use currency::*;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[serde(transparent)]
@@ -99,54 +97,9 @@ pub struct RelayQuery {
     pub mac_commitment: U256,
 }
 
-#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
-pub enum TokenKind {
-    #[default]
-    Native,
-    ERC20(Address),
-}
-
 pub fn server_error(msg: &str) -> Response {
     let code = StatusCode::INTERNAL_SERVER_ERROR;
     (code, SimpleServiceResponse::from(msg)).into_response()
-}
-
-/// A list of all supported coins across all chains. Every relayer instance will work with some
-/// subset of these coins.
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash, EnumIter, Serialize, Deserialize, ValueEnum)]
-pub enum Coin {
-    Eth,
-    Azero,
-    Btc,
-    Usdt,
-    Usdc,
-}
-
-impl Coin {
-    pub fn decimals(&self) -> u32 {
-        match self {
-            Coin::Azero => 18,
-            Coin::Eth => 18,
-            Coin::Btc => 8,
-            Coin::Usdt => 6,
-            Coin::Usdc => 6,
-        }
-    }
-}
-
-impl FromStr for Coin {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "azero" => Ok(Coin::Azero),
-            "eth" => Ok(Coin::Eth),
-            "btc" => Ok(Coin::Btc),
-            "usdt" => Ok(Coin::Usdt),
-            "usdc" => Ok(Coin::Usdc),
-            _ => Err(()),
-        }
-    }
 }
 
 pub const RELATIVE_PRICE_DIGITS: u32 = 20;
